# sis-service/src/main/resources/openapi/api-spec.yaml
openapi: 3.0.3
info:
  title: DCS Digital Wallet API
  description: |
    The DCS Digital Wallet API provides comprehensive student services for the Example University de Humanidades e Tecnologias. 
    This API enables students to access their academic information, manage their digital wallet, and interact with various university services.
  version: 1.0.0
  contact:
    name: DCS Development Team
    email: dev@usis.pt
    url: https://www.usis.pt
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8085/api/v1
    description: Development server
  - url: https://api.usis.pt/v1
    description: Production server
security:
  - studentAuth: []
paths:
...
  /student/login:
    post:
      tags:
        - Authentication
      summary: Student login
      description: |
        Initiates student login and credential workflow process.
        This endpoint starts an asynchronous workflow for student authentication and credential issuance.
        Returns a correlation ID that can be used to track the workflow progress.
      operationId: studentLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentServiceRequest'
      responses:
        '202':
          description: Workflow started successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
              examples:
                processingResponse:
                  summary: Workflow started
                  value:
                    status: "PROCESSING"
                    correlationId: "550e8400-e29b-41d4-a716-446655440000"
                    expiresAt: "2024-01-15T18:30:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /student/registration:
    post:
      tags:
        - Authentication
      summary: Student registration
      description: |
        Handles student registration process with password. 
        This endpoint allows new students to register for the digital wallet service 
        by providing their credentials and platform information.
      operationId: studentRegistration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentRegistrationRequest'
            examples:
              registrationRequest:
                summary: New student registration
                value:
                  userName: "a12345678"
                  language: "PT"
                  platform: "ios"
                  password: "securePassword123"
                  application: "com.example.dcs.mobile"
                  versionCode: "1601206"
      responses:
        '200':
          description: Successfully processed registration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegistrationResponse'
              examples:
                successResponse:
                  summary: Registration successful
                  value:
                    status: "success"
                    message: "Student registered successfully"
                    data:
                      studentId: "a12345678"
                      registrationDate: "2024-01-15T10:30:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    studentAuth:
      type: apiKey
      in: header
      name: X-Student-Auth
      description: Student authentication token
  schemas:
    StudentServiceRequest:
      type: object
      required:
        - userName
        - installKey
        - language
        - platform
        - application
        - versionCode
      properties:
        userName:
          type: string
          description: Unique identifier for the user
          example: "a12345678"
          pattern: "^[a-z0-9]+$"
        installKey:
          type: string
          description: Installation key for the mobile app
          example: "00000_0000000000000"
          minLength: 1
        language:
          type: string
          description: Preferred language of the user
          example: "PT"
          enum: [PT, EN, ES]
        platform:
          type: string
          description: The client platform
          example: "ios"
          enum: [ios, android, web]
        application:
          type: string
          description: Application identifier
          example: "com.example.dcs.mobile"
          minLength: 1
        versionCode:
          type: string
          description: Application version code
          example: "1601206"
          minLength: 1
    StudentRegistrationRequest:
      type: object
      required:
        - userName
        - language
        - platform
        - password
        - application
        - versionCode
      properties:
        userName:
          type: string
          description: Unique identifier for the user
          example: "a12345678"
          pattern: "^[a-z0-9]+$"
        language:
          type: string
          description: Preferred language of the user
          example: "PT"
          enum: [PT, EN, ES]
        platform:
          type: string
          description: The client platform
          example: "ios"
          enum: [ios, android, web]
        password:
          type: string
          description: User's chosen password for authentication
          example: "securePassword123"
          minLength: 8
          format: password
        application:
          type: string
          description: Application identifier
          example: "com.example.dcs.mobile"
          minLength: 1
        versionCode:
          type: string
          description: Application version code
          example: "1601206"
          minLength: 1
...
    LoginResponse:
      type: object
      properties:
        status:
          type: string
          example: "PROCESSING"
          description: "Status of the login request (PROCESSING, COMPLETED, FAILED)"
        correlationId:
          type: string
          example: "550e8400-e29b-41d4-a716-446655440000"
          description: "Correlation ID to track the async workflow progress"
        expiresAt:
          type: string
          format: date-time
          example: "2024-01-15T18:30:00Z"
          description: "Expiration timestamp of the workflow (24 hours from creation)"
    RegistrationResponse:
      type: object
      properties:
        status:
          type: string
          example: "success"
        message:
          type: string
          example: "Student registered successfully"
        data:
          type: object
          properties:
            studentId:
              type: string
              example: "a12345678"
            registrationDate:
              type: string
              format: date-time
              example: "2024-01-15T10:30:00Z"
...
    ErrorResponse:
      type: object
      required:
        - status
        - message
        - errorCode
        - statusCode
        - timestamp
      properties:
        status:
          type: string
          example: "error"
        message:
          type: string
          example: "Validation failed"
        errorCode:
          type: string
          example: "TE-002"
        statusCode:
          type: integer
          example: 400
        timestamp:
          type: string
          format: date-time
        path:
          type: string
          example: "/student/enrolment"
        details:
          type: string
        validationErrors:
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'
    ValidationError:
      type: object
      properties:
        field:
          type: string
          example: "userName"
        message:
          type: string
          example: "userName is required"
        rejectedValue:
          type: string
          example: "null"
  responses:
    BadRequest:
      description: Bad request - invalid input data or missing required fields
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            validationError:
              summary: Missing required field
              value:
                status: "error"
                message: "userName is required"
                errorCode: "TE-002"
                statusCode: 400
                timestamp: "2024-01-15T10:30:00Z"
            invalidData:
              summary: Invalid input data
              value:
                status: "error"
                message: "Invalid platform value. Must be one of: ios, android, web"
                errorCode: "TE-015"
                statusCode: 400
                timestamp: "2024-01-15T10:30:00Z"
    Unauthorized:
      description: Unauthorized - invalid student credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            authenticationError:
              summary: Invalid credentials
              value:
                status: "error"
                message: "Invalid student credentials"
                errorCode: "TE-003"
                statusCode: 401
                timestamp: "2024-01-15T10:30:00Z"
...
    Conflict:
      description: Conflict - student already registered
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            conflictError:
              summary: Student already registered
              value:
                status: "error"
                message: "Student already registered"
                errorCode: "TE-005"
                statusCode: 409
                timestamp: "2024-01-15T10:30:00Z"
    InternalServerError:
      description: Internal server error or external service unavailable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            serverError:
              summary: Internal server error
              value:
                status: "error"
                message: "Unexpected error during operation"
                errorCode: "TE-008"
                statusCode: 500
                timestamp: "2024-01-15T10:30:00Z"
            externalServiceError:
              summary: External service unavailable
              value:
                status: "error"
                message: "SIGES service is currently unavailable"
                errorCode: "TE-013"
                statusCode: 500
                timestamp: "2024-01-15T10:30:00Z"
...